+++ .storybook/componentMap.js
@@ -1,1 +1,3 @@
-module.exports = {};
+import { defaultIcons } from '../packages/styled-components';
+
+export default defaultIcons;