+++ packages/styled-components/components/image/index.js
@@ -1,8 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import sanitizeHtml from 'sanitize-html';
+import useStandardProps from '@irvingjs/styled/hooks/useStandardProps';
+import {
+  standardPropTypes,
+  standardDefaultProps,
+} from '@irvingjs/styled/types/propTypes';
 import { richText } from '@irvingjs/core/config/html';
-import withThemes from '@irvingjs/styled/components/hoc/withThemes';
 import * as defaultStyles from './themes/default';
 
 /**
@@ -33,7 +37,6 @@ const Image = (props) => {
     alt,
     caption,
     children,
-    className,
     credit,
     fallbackSrc,
     height,
@@ -47,7 +50,6 @@ const Image = (props) => {
     theme,
     width,
   } = props;
-
   const {
     FigureWrapper,
     ImageTag,
@@ -56,6 +58,7 @@ const Image = (props) => {
     ImageMeta,
     ImageWrapper,
   } = theme;
+  const standardProps = useStandardProps(props);
 
   /**
    * @todo possibly replace this with similar functionality. This hook breaks b/c it references window.
@@ -84,8 +87,7 @@ const Image = (props) => {
   return (
     <FigureWrapper
       allowUpscaling={allowUpscaling}
-      classsName={className}
-      style={style}
+      {...standardProps}
     >
       <ImageWrapper aspectRatio={aspectRatio}>
         <ImageTag
@@ -120,12 +122,12 @@ const Image = (props) => {
 };
 
 Image.defaultProps = {
+  ...standardDefaultProps,
+  theme: defaultStyles,
   allowUpscaling: false,
   alt: '',
   aspectRatio: false,
   caption: '',
-  children: [],
-  className: '',
   credit: '',
   fallbackSrc: '',
   height: false,
@@ -141,6 +143,7 @@ Image.defaultProps = {
 };
 
 Image.propTypes = {
+  ...standardPropTypes,
   /**
    * Allow an image to be scaled to larger than its actual width.
    */
@@ -165,14 +168,6 @@ Image.propTypes = {
     PropTypes.string,
     PropTypes.bool,
   ]),
-  /**
-   * Children of the component.
-   */
-  children: PropTypes.node,
-  /**
-   * Class name for <ImageWrapper />.
-   */
-  className: PropTypes.string,
   /**
    * Caption.
    */
@@ -238,12 +233,13 @@ Image.propTypes = {
   ]),
 };
 
-export const themeMap = {
+const themeMap = {
   default: defaultStyles,
 };
 
-export { Image as PureComponent };
+export {
+  Image as Component,
+  themeMap,
+};
 
-export const StyledComponent = withThemes(themeMap)(Image);
-
-export default StyledComponent;
+export default Image;