+++ packages/styled-components/components/searchForm/index.js
@@ -3,7 +3,11 @@ import PropTypes from 'prop-types';
 import queryString from 'query-string';
 import { useForm } from 'react-hook-form';
 import history from '@irvingjs/core/utils/history';
-import withThemes from '@irvingjs/styled/components/hoc/withThemes';
+import useStandardProps from '@irvingjs/styled/hooks/useStandardProps';
+import {
+  standardPropTypes,
+  standardDefaultProps,
+} from '@irvingjs/styled/types/propTypes';
 import * as defaultStyles from './themes/default';
 
 /**
@@ -14,12 +18,11 @@ import * as defaultStyles from './themes/default';
 const SearchForm = (props) => {
   const {
     baseUrl,
-    style,
     searchTerm,
     searchTermQueryArg,
-    theme = defaultStyles,
+    theme,
   } = props;
-
+  const standardProps = useStandardProps(props);
   const {
     SearchFormWrapper,
     SearchFormTerm,
@@ -44,7 +47,7 @@ const SearchForm = (props) => {
 
   return (
     <SearchFormWrapper
-      style={style}
+      {...standardProps}
       action={baseUrl}
       onSubmit={handleSubmit(onSubmit)}
     >
@@ -60,14 +63,15 @@ const SearchForm = (props) => {
 };
 
 SearchForm.defaultProps = {
+  ...standardDefaultProps,
+  theme: defaultStyles,
   baseUrl: '/',
   searchTerm: '',
   searchTermQueryArg: 'search',
-  style: {},
-  theme: defaultStyles,
 };
 
 SearchForm.propTypes = {
+  ...standardPropTypes,
   /**
    * Base url for search.
    */
@@ -80,25 +84,15 @@ SearchForm.propTypes = {
    * The query var used for the search term.
    */
   searchTermQueryArg: PropTypes.string,
-  /**
-   * CSS styles.
-   */
-  style: PropTypes.oneOfType([
-    PropTypes.array,
-    PropTypes.object,
-  ]),
-  /**
-   * Theme (styles) to apply to the component.
-   */
-  theme: PropTypes.object,
 };
 
-export const themeMap = {
+const themeMap = {
   default: defaultStyles,
 };
 
-export { SearchForm as PureComponent };
+export {
+  SearchForm as Component,
+  themeMap,
+};
 
-export const StyledComponent = withThemes(themeMap)(SearchForm);
-
-export default StyledComponent;
+export default SearchForm;