+++ packages/styled-components/components/socialSharing/index.js
@@ -2,24 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { __, sprintf } from '@wordpress/i18n';
 import queryString from 'query-string';
-import withThemes from '@irvingjs/styled/components/hoc/withThemes';
-import SocialSharingItem from './socialSharingItem';
+import useStandardProps from '@irvingjs/styled/hooks/useStandardProps';
+import {
+  standardPropTypes,
+  standardDefaultProps,
+} from '@irvingjs/styled/types/propTypes';
+import toReactElement from '@irvingjs/core/utils/toReactElement';
+import Link from '../link';
 import * as defaultStyles from './themes/default';
-
-/**
- * Supported platforms.
- *
- * @type {Array}
- */
-export const SupportedPlatforms = [
-  'email',
-  'facebook',
-  'linkedin',
-  'pinterest',
-  'reddit',
-  'twitter',
-  'whatsapp',
-];
 
 /**
  * Social sharing.
@@ -31,13 +21,15 @@ const SocialSharing = (props) => {
     description,
     imageUrl,
     platforms,
-    style,
+    platformShareLinks,
     theme,
     title,
     url,
   } = props;
-
+  const standardProps = useStandardProps(props);
   const {
+    IconWrapper,
+    SocialSharingItemWrapper,
     SocialSharingList,
     SocialSharingWrapper,
   } = theme;
@@ -107,36 +99,68 @@ const SocialSharing = (props) => {
     whatsapp: getWhatsAppUrl,
   };
 
+  const getShareUrl = (platform) => {
+    if (platformShareLinks[platform]) {
+      return platformShareLinks[platform];
+    }
+
+    if (socialUrlMap[platform]) {
+      return socialUrlMap[platform];
+    }
+
+    return '#';
+  };
+
+  if (0 === platforms.length) {
+    return null;
+  }
+
+  const items = platforms.map((platform) => ({
+    platform,
+    shareUrl: getShareUrl(platform),
+    icon: toReactElement({
+      name: `irving/${platform}-icon`,
+      config: {
+        title: sprintf( // Translators: %1$s - platform.
+          __('Share on %1$s', 'irving-styled-components'),
+          platform
+        ),
+      },
+      children: [],
+    }),
+  }));
+
   return (
-    <SocialSharingWrapper style={style}>
-      {platforms && 0 !== platforms.length && (
-        <SocialSharingList>
-          {platforms.map((platform) => (
-            <SocialSharingItem
-              key={platform}
-              platform={platform}
-              theme={theme}
-              title={platform}
-              url={socialUrlMap[platform]}
-            />
-          ))}
-        </SocialSharingList>
-      )}
+    <SocialSharingWrapper {...standardProps}>
+      <SocialSharingList>
+        {items.map(({ platform, shareUrl, icon }) => (
+          <SocialSharingItemWrapper key={platform} className={platform}>
+            <Link href={shareUrl}>
+              <IconWrapper className={platform}>
+                {icon}
+              </IconWrapper>
+            </Link>
+          </SocialSharingItemWrapper>
+        ))}
+      </SocialSharingList>
     </SocialSharingWrapper>
   );
 };
 
 SocialSharing.defaultProps = {
+  ...standardDefaultProps,
+  theme: defaultStyles,
   description: '',
   imageUrl: '',
   platforms: ['email', 'facebook', 'twitter'],
+  platformShareLinks: {},
   style: {},
-  theme: defaultStyles,
   title: '',
   url: '',
 };
 
 SocialSharing.propTypes = {
+  ...standardPropTypes,
   /**
    * Description of the shared content.
    */
@@ -148,18 +172,14 @@ SocialSharing.propTypes = {
   /**
    * An array of social sharing platforms.
    */
-  platforms: PropTypes.arrayOf(PropTypes.oneOf(SupportedPlatforms)),
+  platforms: PropTypes.arrayOf(PropTypes.string),
   /**
-   * CSS styles.
+   * An object containing social platforms as keys and share links as values.
    */
-  style: PropTypes.oneOfType([
+  platformShareLinks: PropTypes.oneOfType([
     PropTypes.array,
     PropTypes.object,
   ]),
-  /**
-   * Theme (styles) to apply to the component.
-   */
-  theme: PropTypes.object,
   /**
    * Title of the shared content.
    */
@@ -170,12 +190,13 @@ SocialSharing.propTypes = {
   url: PropTypes.string,
 };
 
-export const themeMap = {
+const themeMap = {
   default: defaultStyles,
 };
 
-export { SocialSharing as PureComponent };
+export {
+  SocialSharing as Component,
+  themeMap,
+};
 
-export const StyledComponent = withThemes(themeMap)(SocialSharing);
-
-export default StyledComponent;
+export default SocialSharing;