+++ CHANGELOG.md
@@ -3,6 +3,156 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0-alpha.0](https://github.com/alleyinteractive/irving/compare/v3.3.0...v3.4.0-alpha.0) (2020-09-01)
+
+
+
+# [3.3.0-beta.6](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.5...v3.3.0-beta.6) (2020-08-31)
+
+
+### Bug Fixes
+
+* **sc:** publish types ([af49ebb](https://github.com/alleyinteractive/irving/commit/af49ebbc7c4978ce789660bc130a7ee6199e2916))
+
+
+
+# [3.3.0-beta.5](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.4...v3.3.0-beta.5) (2020-08-31)
+
+
+### Bug Fixes
+
+* **package:** publsih sc componentmap ([e519ffd](https://github.com/alleyinteractive/irving/commit/e519ffd20f7685fbd0613337cf42da2a4874fcb1))
+
+
+
+# [3.3.0-beta.4](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.3...v3.3.0-beta.4) (2020-08-31)
+
+
+### Bug Fixes
+
+* **styledcomponents:** address CR feedback ([ec1694a](https://github.com/alleyinteractive/irving/commit/ec1694a89cc3e5d84134e757b64e58fd7bab227b))
+
+
+### Features
+
+* **<head> and <helmet> management:** ensures that style and script tags work well ([5558442](https://github.com/alleyinteractive/irving/commit/5558442f867750a6598dc2b7bb82f191907b10aa))
+
+
+
+# [3.3.0-beta.3](https://github.com/alleyinteractive/irving/compare/v3.3.0-rc.0...v3.3.0-beta.3) (2020-08-31)
+
+
+### Bug Fixes
+
+* **getstandardprops:** fix some issues with proptypes ([72c33e2](https://github.com/alleyinteractive/irving/commit/72c33e26f237ba331103749e90d44472d9cdec27))
+* **package:** fix mismatched hoisted package versions ([f75ac82](https://github.com/alleyinteractive/irving/commit/f75ac82a2d34b276e1b00eb89f37dd8008830ecd))
+* **sociallinks:** fix icon display issue, add class to icon wrapper, and tweak CSS for social links ([132f463](https://github.com/alleyinteractive/irving/commit/132f46320224618b087305f95fe9782a9f6f9382))
+
+
+### Features
+
+* **socialsharing:** allow for custom links in Social Sharing component and refactor to one file ([54f36b3](https://github.com/alleyinteractive/irving/commit/54f36b36f330876540789bbaf32725136765403d))
+* **styledcomponents:** add new standard prop types, default props, props values to sc package ([204d824](https://github.com/alleyinteractive/irving/commit/204d82478061810802be22b8a41b4084a138fbab))
+* **withthemes:** refactor sc package to remove now-unnecessary createWithUserthemes hoc ([2b475ce](https://github.com/alleyinteractive/irving/commit/2b475cedf2acca429b0b9e648fc19e2233af379f))
+
+
+
+
+
+# [3.3.0](https://github.com/alleyinteractive/irving/compare/v3.3.0-rc.0...v3.3.0) (2020-09-01)
+
+**Note:** Version bump only for package irving
+
+
+# [3.3.0-rc.0](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.2...v3.3.0-rc.0) (2020-08-31)
+
+**Note:** Version bump only for package irving
+
+# [3.3.0-beta.6](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.5...v3.3.0-beta.6) (2020-08-31)
+
+
+### Bug Fixes
+
+* **sc:** publish types ([af49ebb](https://github.com/alleyinteractive/irving/commit/af49ebbc7c4978ce789660bc130a7ee6199e2916))
+
+
+
+
+
+# [3.3.0-beta.5](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.4...v3.3.0-beta.5) (2020-08-31)
+
+
+### Bug Fixes
+
+* **package:** publsih sc componentmap ([e519ffd](https://github.com/alleyinteractive/irving/commit/e519ffd20f7685fbd0613337cf42da2a4874fcb1))
+
+
+
+
+
+# [3.3.0-beta.4](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.3...v3.3.0-beta.4) (2020-08-31)
+
+
+### Bug Fixes
+
+* **getstandardprops:** fix some issues with proptypes ([72c33e2](https://github.com/alleyinteractive/irving/commit/72c33e26f237ba331103749e90d44472d9cdec27))
+* **package:** fix mismatched hoisted package versions ([f75ac82](https://github.com/alleyinteractive/irving/commit/f75ac82a2d34b276e1b00eb89f37dd8008830ecd))
+* **styledcomponents:** address CR feedback ([ec1694a](https://github.com/alleyinteractive/irving/commit/ec1694a89cc3e5d84134e757b64e58fd7bab227b))
+
+
+### Features
+
+* **styledcomponents:** add new standard prop types, default props, props values to sc package ([204d824](https://github.com/alleyinteractive/irving/commit/204d82478061810802be22b8a41b4084a138fbab))
+* **withthemes:** refactor sc package to remove now-unnecessary createWithUserthemes hoc ([2b475ce](https://github.com/alleyinteractive/irving/commit/2b475cedf2acca429b0b9e648fc19e2233af379f))
+
+
+
+
+
+# [3.3.0-beta.3](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.2...v3.3.0-beta.3) (2020-08-31)
+
+
+### Bug Fixes
+
+* **sociallinks:** fix icon display issue, add class to icon wrapper, and tweak CSS for social links ([132f463](https://github.com/alleyinteractive/irving/commit/132f46320224618b087305f95fe9782a9f6f9382))
+
+
+### Features
+
+* **socialsharing:** allow for custom links in Social Sharing component and refactor to one file ([54f36b3](https://github.com/alleyinteractive/irving/commit/54f36b36f330876540789bbaf32725136765403d))
+
+
+# [3.3.0-beta.2](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.1...v3.3.0-beta.2) (2020-08-27)
+
+
+### Bug Fixes
+
+* **commentcount:** fix reference ([3c5e554](https://github.com/alleyinteractive/irving/commit/3c5e5546730cfb0e2a5171931bbbb46ae5ed7c6f))
+
+
+
+
+
+# [3.3.0-beta.1](https://github.com/alleyinteractive/irving/compare/v3.3.0-beta.0...v3.3.0-beta.1) (2020-08-27)
+
+
+### Bug Fixes
+
+* **cacheclient:** return default client, not service, from VIP go cacheClient ([07a2249](https://github.com/alleyinteractive/irving/commit/07a2249db017f9c998b11e0cccd73497681738d4))
+* **Coral:** Fix client rendering activation of comment counts ([3a48963](https://github.com/alleyinteractive/irving/commit/3a489633f1fea785ccfe32f75b1e11057636a97b))
+
+
+
+
+
+# [3.3.0-beta.0](https://github.com/alleyinteractive/irving/compare/v3.3.0-alpha.1...v3.3.0-beta.0) (2020-08-27)
+
+**Note:** Version bump only for package irving
+
+
+
+
+
 ## [3.2.1](https://github.com/alleyinteractive/irving/compare/v3.2.1-alpha.0...v3.2.1) (2020-08-21)