+++ packages/wp-admin-bar/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0-alpha.0](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.3.0...v3.4.0-alpha.0) (2020-09-01)
+
+**Note:** Version bump only for package @irvingjs/wp-admin-bar
+
+
+
+
+
+# [3.3.0](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.3.0-rc.0...v3.3.0) (2020-09-01)
+
+**Note:** Version bump only for package @irvingjs/wp-admin-bar
+
+
+
+
+
+# [3.3.0-rc.0](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.3.0-beta.2...v3.3.0-rc.0) (2020-08-31)
+
+**Note:** Version bump only for package @irvingjs/wp-admin-bar
+
+
+
+
+
+# [3.3.0-beta.1](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.3.0-beta.0...v3.3.0-beta.1) (2020-08-27)
+
+**Note:** Version bump only for package @irvingjs/wp-admin-bar
+
+
+
+
+
+# [3.3.0-beta.0](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.3.0-alpha.1...v3.3.0-beta.0) (2020-08-27)
+
+**Note:** Version bump only for package @irvingjs/wp-admin-bar
+
+
+
+
+
 # [3.2.0](https://github.com/alleyinteractive/irving/packages/wp-admin-bar/compare/v3.2.0-beta.1...v3.2.0) (2020-08-12)
 
 **Note:** Version bump only for package @irvingjs/wp-admin-bar